Agregando algo xd

+++ src/main/java/com/udec/myfinancialbook/repository/IBookDayRepo.java
@@ -3,6 +3,6 @@ package com.udec.myfinancialbook.repository;
 import com.udec.myfinancialbook.model.BookDay;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IBookDayRepo extends JpaRepository<BookDay, Integer> {
+public interface IBookDayRepo extends JpaRepository<BookDay, Long> {
 
 }
